+++ .eslintrc.js
@@ -6,19 +6,28 @@ module.exports = {
     node: true,
   },
   parserOptions: {
-    parser: 'babel-eslint',
+    parser: '@typescript-eslint/parser',
+    ecmaVersion: 2020,
+    sourceType: 'module',
+    lib: ['es2020'],
+    ecmaFeatures: {
+      jsx: true,
+      tsx: true,
+    },
+    extraFileExtensions: ['.vue'],
   },
+  plugins: ['vue', 'jsdoc', '@typescript-eslint', 'prettier'],
   extends: [
     'plugin:vue-libs/recommended',
     'plugin:prettier/recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:jsdoc/recommended',
     'prettier',
-    'prettier/vue',
   ],
-  // required to lint *.vue files
-  plugins: ['vue', 'prettier'],
   // add your custom rules here
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'vue/require-render-return': 'off',
   },
 };