+++ types/index.d.ts
@@ -1,0 +1,15 @@
+import Vue, { VueConstructor } from 'vue';
+
+export const MglMap: VueConstructor<Vue>;
+export const MglNavigationControl: VueConstructor<Vue>;
+export const MglGeolocateControl: VueConstructor<Vue>;
+export const MglFullscreenControl: VueConstructor<Vue>;
+export const MglAttributionControl: VueConstructor<Vue>;
+export const MglScaleControl: VueConstructor<Vue>;
+export const MglGeojsonLayer: VueConstructor<Vue>;
+export const MglImageLayer: VueConstructor<Vue>;
+export const MglVideoLayer: VueConstructor<Vue>;
+export const MglVectorLayer: VueConstructor<Vue>;
+export const MglRasterLayer: VueConstructor<Vue>;
+export const MglMarker: VueConstructor<Vue>;
+export const MglPopup: VueConstructor<Vue>;